+++ CHANGES.md
@@ -15,7 +15,9 @@
     the features in the dataset that intersect the bbox.
 -   Addition of `set_gdal_config_options` to set GDAL configuration options and
     `get_gdal_config_option` to get a GDAL configuration option.
--   Addition of `pyogrio.__gdal_version__` to return GDAL version tuple.
+-   Addition of `pyogrio.__gdal_version__` attribute to return GDAL version tuple
+    and `__gdal_version_string__` to return string version.
+-   Addition of `list_drivers` function to list all available GDAL drivers.
 
 ## 0.1.0